Move ground detection into own plugin+module
