+++ examples/platformer/player.rs
@@ -2,7 +2,7 @@ use bevy::prelude::*;
 use bevy_ecs_ldtk::prelude::*;
 use bevy_rapier2d::dynamics::Velocity;
 
-use crate::physics::{ColliderBundle, GroundDetection};
+use crate::{ground_detection::GroundDetection, physics::ColliderBundle};
 use crate::{climbing::Climber, inventory::Inventory};
 
 #[derive(Copy, Clone, Eq, PartialEq, Debug, Default, Component)]
